The worst attempt at making the graphing stuff work
So the idea for the inputs is that this program will be a friend user interface right. So like, instead of going into the code you will be presented with a GUI to enter the right inputs. For now, there is no fancy GUI so user will input the values into terminal.